requirements.txt
Django==1.8.7
django-filter==0.11.0
djangorestframework==3.3.1
wheel==0.24.0
